Major API change to support progress and cancellation

+++ MStorage/FilesystemStorage/FilesystemStorage.cs
@@ -1,8 +1,10 @@
-﻿using System;
+﻿using HttpProgress;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.IO.Enumeration;
 using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace MStorage.FilesystemStorage
@@ -30,8 +32,12 @@ namespace MStorage.FilesystemStorage
         /// Deletes the entire root directory including all stored files and subdirectories.
         /// This is a dangerous operation if the root directory is shared!
         /// </summary>
-        public Task DeleteAllAsync()
+        /// <param name="progress">This event is not supported on this backedn and so will never fire.</param>
+        /// <param name="cancel">Allows cancellation of the delete operation.</param>
+        public Task DeleteAllAsync(IProgress<long> progress = null, CancellationToken cancel = default(CancellationToken))
         {
+            if (cancel.IsCancellationRequested) { return Task.FromCanceled(cancel); }
+
             try
             {
                 Directory.Delete(RootDirectory, true);
@@ -50,19 +56,39 @@ namespace MStorage.FilesystemStorage
         /// Deletes a single file from disk by the given name. Throws FileNotFound exception if it doesnt exist.
         /// </summary>
         /// <param name="name">The file to delete.</param>
-        public Task DeleteAsync(string name)
+        /// <param name="cancel">Allows cancellation of the delete operation.</param>
+        public Task DeleteAsync(string name, CancellationToken cancel = default(CancellationToken))
         {
+            if (cancel.IsCancellationRequested) { return Task.FromCanceled(cancel); }
+
             File.Delete(GetFullPath(name));
             return Task.CompletedTask;
         }
 
         /// <summary>
-        /// Calls OpenRead on the given object name.
+        /// Calls OpenRead on the given object name, returning a FileStream. Throws FileNotFound if the object does not exist.
         /// </summary>
-        /// <param name="name">The object name to open.</param>
-        /// <returns></returns>
-        public Task<Stream> DownloadAsync(string name)
+        /// <param name="name">The name of the object to retrieve.</param>
+        /// <param name="progress">Progress events are not fired for this implementation.</param>
+        /// <param name="cancel">Allows cancellation of the transfer.</param>
+        /// <returns>A stream containing the requested object. The underlying type is FileStream.</returns>
+        public Task<Stream> DownloadAsync(string name, CancellationToken cancel = default(CancellationToken))
         {
+            if (cancel.IsCancellationRequested) { return Task.FromCanceled<Stream>(cancel); }
+            return Task.FromResult<Stream>(File.OpenRead(GetFullPath(name)));
+        }
+
+        /// <summary>
+        /// Calls OpenRead on the given object name, returning a FileStream. Throws FileNotFound if the object does not exist.
+        /// </summary>
+        /// <param name="name">The name of the object to retrieve.</param>
+        /// <param name="output">The output stream data will be copied to.</param>
+        /// <param name="progress">Progress events are not fired for this implementation.</param>
+        /// <param name="cancel">Allows cancellation of the transfer.</param>
+        /// <returns>A stream containing the requested object. The underlying type is FileStream.</returns>
+        public Task DownloadAsync(string name, Stream output, IProgress<ICopyProgress> progress = null, CancellationToken cancel = default(CancellationToken))
+        {
+            if (cancel.IsCancellationRequested) { return Task.FromCanceled<Stream>(cancel); }
             return Task.FromResult<Stream>(File.OpenRead(GetFullPath(name)));
         }
 
@@ -70,9 +96,12 @@ namespace MStorage.FilesystemStorage
         /// Retrieves a list of all files stored in the root directory.
         /// Subdirectories and inaccessible files are not returned.
         /// </summary>
+        /// <param name="cancel">Allows cancellation of the list operation.</param>
         /// <returns>A collection of object names.</returns>
-        public Task<IEnumerable<string>> ListAsync()
+        public Task<IEnumerable<string>> ListAsync(CancellationToken cancel = default(CancellationToken))
         {
+            if (cancel.IsCancellationRequested) { return Task.FromCanceled<IEnumerable<string>>(cancel); }
+
             return Task.FromResult(ListFiles());
         }
 
@@ -82,36 +111,39 @@ namespace MStorage.FilesystemStorage
         /// <param name="destination">The instance to transfer to.</param>
         /// <param name="deleteSource">Delete each object in this store after it has successfully been transferred.</param>
         /// <returns>A collection of statuses indicating the success or failure state for each transfered object.</returns>
-        public async Task<IEnumerable<StatusedValue<string>>> TransferAsync(IStorage destination, bool deleteSource)
+        public async Task TransferAsync(IStorage destination, bool deleteSource, IProgress<string> success = null, IProgress<ExceptionWithValue<string>> error = null, CancellationToken cancel = default(CancellationToken))
         {
+            if (cancel.IsCancellationRequested) { return; }
+
             if (Equals(destination))
             {
                 // Target is same as source. Nothing to do.
-                return Enumerable.Empty<StatusedValue<string>>();
+                return;
             }
 
-            var result = new List<StatusedValue<string>>();
+            var result = new List<ExceptionWithValue<string>>();
             foreach (var filename in ListFiles())
             {
+                if (cancel.IsCancellationRequested) { return; }
+
                 try
                 {
                     string fullPath = GetFullPath(filename);
                     using (var fileStream = File.OpenRead(fullPath))
                     {
-                        await destination.UploadAsync(filename, fileStream);
+                        await destination.UploadAsync(filename, fileStream, cancel: cancel);
                     }
                     if (deleteSource)
                     {
                         File.Delete(fullPath);
                     }
-                    result.Add(new StatusedValue<string>(filename, true, null));
+                    if (success != null) { success.Report(filename); }
                 }
                 catch (Exception ex)
                 {
-                    result.Add(new StatusedValue<string>(filename, false, ex));
+                    if (error != null) { error.Report(new ExceptionWithValue<string>(filename, ex)); }
                 }
             }
-            return result;
         }
 
         /// <summary>
@@ -120,13 +152,17 @@ namespace MStorage.FilesystemStorage
         /// <param name="name">The filename to give this object.</param>
         /// <param name="file">The stream to upload.</param>
         /// <param name="disposeStream">If true, the file stream will be closed automatically after being consumed.</param>
-        public async Task UploadAsync(string name, Stream file, bool disposeStream)
+        /// <param name="progress">Fires periodically with transfer progress.</param>
+        /// <param name="cancel">Allows cancellation of the transfer.</param>
+        /// <param name="expectedStreamLength">Allows overriding the stream's expected length for progress reporting as some stream types do not support Length.</param>
+        public async Task UploadAsync(string name, Stream file, bool disposeStream = false, IProgress<ICopyProgress> progress = null, CancellationToken cancel = default(CancellationToken), long expectedStreamLength = 0)
         {
+            long streamLength = file.CanSeek && expectedStreamLength == 0 ? file.Length : expectedStreamLength;
             try
             {
                 using (var fileStream = File.Open(GetFullPath(name), FileMode.Create))
                 {
-                    await file.CopyToAsync(fileStream);
+                    await file.CopyToAsync(fileStream, expectedTotalBytes: streamLength, progressReport: progress, cancelToken: cancel);
                 }
             }
             finally
@@ -141,17 +177,22 @@ namespace MStorage.FilesystemStorage
         /// <param name="name">The filename to give this object.</param>
         /// <param name="path">A path to the file to upload.</param>
         /// <param name="deleteSource">If true, the file on disk will be deleted after the upload is complete.</param>
-        public Task UploadAsync(string name, string path, bool deleteSource)
+        /// <param name="progress">Fires periodically with transfer progress.</param>
+        /// <param name="cancel">Allows cancellation of the transfer.</param>
+        public async Task UploadAsync(string name, string path, bool deleteSource, IProgress<ICopyProgress> progress = null, CancellationToken cancel = default(CancellationToken))
         {
+            if (cancel.IsCancellationRequested) { return; }
+
             string destFileName = GetFullPath(name);
+            var info = new FileInfo(path);
             if (destFileName != path)
             {
-                if (File.Exists(destFileName)) { File.Delete(destFileName); }
-                File.Copy(path, destFileName);
+                using (var s = File.OpenRead(path))
+                {
+                    await UploadAsync(name, s, deleteSource, progress, cancel, info.Length);
+                }
                 if (deleteSource) { File.Delete(path); }
             }
-
-            return Task.CompletedTask;
         }
 
         private string GetFullPath(string filename)
